requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
biopython==1.81
openai>=1.0.0
numpy==1.24.3